requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
email-validator
APScheduler
gunicorn
